pull util functions out of Commands classes
moved project functions to Server class,  moved user functions into UserData class
