+++ tabcmd/commands/project/publish_samples_command.py
@@ -1,9 +1,11 @@
 from tabcmd.commands.auth.session import Session
 from tabcmd.execution.logger_config import log
-from .project_command import *
+from tabcmd.commands.auth.session import Session
+from tabcmd.commands.server import Server
+from tabcmd.execution.logger_config import log
 
 
-class PublishSamplesCommand(ProjectCommand):
+class PublishSamplesCommand(Server):
     """
     Command to Publish Tableau Sample workbooks to the specified project.
     Any existing samples will be overwritten.
@@ -19,7 +21,7 @@ class PublishSamplesCommand(ProjectCommand):
         session = Session()
         server = session.create_session(args)
         if args.parent_path_name is not None:
-            project_path = ProjectCommand.find_project_id(server, args.parent_path_name)
+            project_path = Server.find_project_id(server, args.parent_path_name)
         else:
             project_path = None
         Server.exit_with_error(logger, "Not yet implemented")
