+++ tabcmd/commands/user/remove_users_command.py
@@ -3,7 +3,7 @@ import tableauserverclient as TSC
 from tabcmd.commands.auth.session import Session
 from tabcmd.commands.constants import Errors
 from tabcmd.execution.logger_config import log
-from .user_command import UserCommand
+from .user_data import UserCommand
 
 
 class RemoveUserCommand(UserCommand):
